requirements.txt
# Core dependencies
# Python 3.9+ required

# LLM Integration
ollama>=0.1.0
transformers>=4.35.0
torch>=2.0.0
numpy<2.0.0
requests>=2.31.0

# Testing frameworks
pytest>=7.4.0
pytest-cov>=4.1.0
mutmut>=2.4.0
coverage>=7.3.0

# Utilities
pyyaml>=6.0.1
click>=8.1.7
tqdm>=4.66.0
colorama>=0.4.6

# Reporting
jinja2>=3.1.2
matplotlib>=3.8.0
pandas>=2.1.0

